fix: incorrectly checking for proper modifier.
Because the length attribute was included in the check with a
conditional AND, as-well-as the length attribute being present would
cause this check to always be false and thus we run into an issue where
we try to create a password with no modifications. This inturn would
cause a recursion error later on.

Because the length attribute is already being checked for separately we
do not need to include it here.

Signed-off-by: staticBanter <staticbanter@staticblogs.ca>