fix: incorrect check of modifier attributes.
simplePass preforms a check to determine the number of set modifier
attributes that should be applied to the password. This is to prevent
too many modifiers being applied to a password that is too short. The
problem was simplePass didn't check if the attributes are actually valid
simplePass attributes and thus could be incorrect in the number of set
attributes. This has been correct by preforming the following:
* creating a list of allowed modifier attributes.
* preforming a check to determine if the set attribute is included in
the list

This update also moves this attribute checking step out of the
`conformPassword()` function and directly into the `simplePass()` main
function call as it is more logical to only preform this check once.

Update JSDoc comments to reflect changes.

Signed-off-by: staticBanter <staticbanter@staticblogs.ca>