refactor: cleaned up redundant `if` statements.
Replaced what used to be individual `if` checks with a loop.

This loop iterates over a object that contains properties that are the
same as a `passwordModifier` object. The values of these properties are
the Regular Expression checks that need to be preformed on the password
to check for conformity.

Other than this the function performs the exact same as before.

Signed-off-by: staticBanter <staticbanter@staticblogs.ca>