docs: removed manuel version numbering
Though these numbers (should) provide good visual cues as to what
version of the software you are using, having to update them manually
was cumbersome and could be error pron.

So until an more streamline approach for this problem is implemented the
only versioning will be done via GitHub (Release) tags and in the
`package.json`.

Signed-off-by: staticBanter <staticbanter@staticblogs.ca>