fix: ending whitespace characters could be replaced.
`conformPassword()` would sometimes replace the ending whitespace
character of a password even when `w_ending_required` was set. This was
due to the `slicePoint` being allowed to choose a position in the
password that may cut off the ending character.

Fixed by checking if either `w_begining_required` or `w_end_required`
are set. This is similar to how the `w_between` modifier worked where if
the password was missing a whitespace character in the middle of the
password `slicePoint` must be a position in the middle to ensure the new
whitespace character will be inserted properly.

Signed-off-by: staticBanter <staticbanter@staticblogs.ca>