feat: can now accept a `FormData` object
simplePass will now take a `FormData` object and use it like normal as
long as the inputs follow standard Javascript conventions and use the
required modifier names. To implement this feature we need to do a few
things, such as:

* Updating the demo code to use the new method.
* Creating a new helper function `cleanModifier`, to *clean*/convert the
modifier object. This also moved some of the code that was providing
this functionality in the base `simplePass` function into here.
* The length attribute is now optional, but will still be required (this
is to allow the `cleanModifier` function to initiate a blank
`I_PasswordModifier` without having to redefine the length attribute.).

Signed-off-by: staticBanter <staticbanter@staticblogs.ca>