refactor: changed how characters are generated.
This update changes how the `generateCharCode` function generates
characters. The new method will allow for the expansion of more
characters by removing the manual `if else` chaining. Furthermore we
also removed the `conformPassword` check step as the new method will
always ensure the proper number of attributes exists in the password (at
the cost of possibly being a bit less random).

This update includes the following changes:
* The `generateCharCode` function now recieves a new parameter, a
`charCodeRequest` object which contains a string that is the desired
character we want to generate. The checks are now performed against a
`characterCodeConstraints` object which contains the attributes for each
character type.
* Removed the `conformPassword` function.
* Moved all Interfaces, Enums, Lists and Objects into a Data folder.
* Removed any currently unused modifiers such as, *memorable* and
*special*
* Updated the README
* Updated the Demo page:
	* Clarified the character set being used.
	* Added a warning about excessive use of the *exclude characters*
feature
	* Removed unneeded JavaScript code for the demo
* Updates to the `package.json` file:
	* removed the `main` attribute, replaced with `browser`
	* Updated version.
	* Added more directories.
	* Changed the `license` attribute to `licenses` and added a URL.
* Added two new helper functions:
	* `shuffle` used to shuffle arrays.
	* `createModifierList` creates a list of modifiers from an object.

Signed-off-by: staticBanter <staticbanter@staticblogs.ca>