docs: reformatted documentation.
The documentation has now been reformatted so that it can now be
displayed both locally and via the GitHub Pages. This has been achieved
through the following:
1. Moving the output of JSDocs files to a separate directory called
```site```. This also resulted in changing the
```workflows/static.yml``` to point to the new directory.
2. Moved all markdown documentation related files into the ```docs```
directory.
3. All links with in the markdown documentation was converted to
relative paths within the ```docs``` directory and removed all file
extensions.
4. Updated the ```docs-build``` changing the ```tutorials``` option to
point the new markdown location.

Signed-off-by: staticBanter <staticbanter@staticblogs.ca>