docs: Restructured the documentation & updated the documentation.
* The directory for the template files used to generate the
documentation site has been renamed to ‘site_template’ for clarity.
* Moved  the README back to the root directory.

This also resulted in a slight change to the ```jsdoc.config.json```
file.

* Enabled ID attributes for site documentation headings.
* Disabled time-stamp generation for documentation.
* Removed documentation referencing between markdown files. This was
causing to many problems when trying to generate the site with JSDoc.
Also in ‘Development’ removed the explanation about how to link markdown
files.
* Added note in ‘Installation in Development’ about needing to run
simplePass in server environment.
* Fixed some typos in the TypeScript examples.
* Updated the CONTRIBUTING file. New methods of contributing!
* Added a little thank you to the bottom of each documentation and a
reminder that they can go to GitHub to open an Issue if needed.
* Added a ‘Refactor Request’ template file.

Signed-off-by: staticBanter <staticbanter@staticblogs.ca>