Closes #8: Ability To Request Repeatable Characters.
Added the ability to Ensure Repeatable Characters in the password. The
following can also be achieved:
* Ensuring "x" amount of repeating characters in the password.
* Ensuring "x" amount of custom repeating characters in the password.

* A new function called `ensureRepeatingCharacters` that will return a
string that contains repeating characters. This function also brought
with it a new Interface to describe its options object.

* Updates to the demo to display the new feature.
* Validation logic updated to account for the new attributes that have
been added.
* Core ```simplePass``` function had to be updated to account for new
feature.
* Modifier list has been updated to list new attributes that are allowed
on the password modifier, as well as the password modifier interface
being updated with new attributes.
* Replace ```whitespaceBetweenMissingLimit``` error with
```missingRequiredAttriubte``` as this explains what is actually
happening and can apply to more situations.
* Updated the docs and changed the site directory structure so it is
more consistent in the future.

Signed-off-by: staticBanter <staticbanter@staticblogs.ca>