Docs: Changed the name of the demo page.
Changed the name of the demo page from 'demo' to 'index' this is because
we are going to be using it as the index page for the GitHub Pages.

Changed the pathing to the GitHub Pages in the 'workflows' file to the
root directory rather than the JsDocs generated folder.

Signed-off-by: staticBanter <staticbanter@staticblogs.ca>