Closes: #38. PWA Support
This commit brings basic PWA functionality to the program including:

* The ability to add to home screen.
* Work offline.
* Discoverable (uses Open Graph Protocol meta tags).
* Linkable.

This involved having to add a service worker ```sw.js``` and a
```manifest.json``` to the project as well as some simple icons (i tried
to keep them as small as i could).

With this commit we also move all the 'demo' code into the site repo to
keep everything clean. We have also renamed the demo asset files to
```main.js``` and ```main.css```

We have renamed the site docs directory from 'simplepass' to 'docs' for
simplicity, because of this we had to modify the ```docs-build```
command. And we have added a new tool to the development pipeline to
shrink the size of the doc files outputted from JSDoc called
'html-minifier'. This helps make the PWA bundle smaller. With this tool
we have also included:

* A new config file ```html-minifier.config.json```
* A new command *NPM* command ```npm run html-min``` - Minifies the html
docs. This command has been added to the ```production``` command chain
aswell.

We have added a new Webpack command ```webpack-watch```. This is because
the bundled version of the program is required for the PWA.

We have also managed to sneak-in another new feature as-well!

Support for Basic Latin(1) Supplement characters has now been added!