!Closes: #14 Added a strength checker.
- Add functionality to preform a strength check on a password. The main
thing being the ability to calculate the binary entropy of the password.
This included some new interfaces, the ```strengthScores``` object, and
three new functions:
	- ```strengthChecker``` - checks the password strength
	- ```characterStrengthChecker``` - checks if the password contains
enough characters and good mix of unique and random characters.
- ```calculateMaxPossibleCharacters``` - calculates the maximum number
of character for a give number of character-sets.

With this new functionality simplePass may now also visual style an
element based on the password strength. See the documentation for more
details. It is also completely optional to use.

- ! Changed up the naming conventions for a lot of files and directories
in an effort to make things more straight forward.
- SASS is now used to help build the U.I for the demo.
- Close: #44 By correcting the logical issue.
- Update the default password configuration to be stronger.
- ! Some configuration variables had their names changed to conform to
the style guide.

Signed-off-by: staticBanter <staticbanter@staticblogs.ca>