feat: Added "Latin Extended A" character support.
- feat: Added support for "Latin Extended A"
- feat: The inclusion of this character-set also came with a new
function, "range". This is a very primitive range function but it does
the trick. It is needed to create a list of alternating characters and
exclude some specific numbers along the way. We use this function to
generate the new character-sets range for certain character types.
- feat: ```characterCodeConstraintsAttributes``` now has a new property,
"offset".
- docs: Added missing property values on the
```characterCodeConstraints``` object.
- fix: U.I used different rounding function on the first password
generated resulting in a different entropy value being shown.

Signed-off-by: staticBanter <staticbanter@staticblogs.ca>