rel: Release 9.0.0-Beta
* Use of Webpacks Banner plugin to ensure license banner is included in
the bundle.
* Use of Webpacks Terser plugin to ensure separate ```license.txt```
files are not created.

* License comment banner has been add to all source files. This is to
comply with the "How to Apply These Terms to Your New Programs" section
of the GNU GENERAL PUBLIC LICENSE.
* ! The ```displayMessages``` parameter of the ```simplePass()```
function has been removed. It has instead been replaced by incorporating
its attributes within the ```config.messages``` object.
* Removed old used *requires* in documentation for various functions.
* ! Cleaned up the ```config``` object. This is to better distinguish
what each setting actually does. This also resulted in any functions
using this object to have to be refactored as well.
* Initializer has been updated to be more easily customizable using the
configuration object.
* The 'EXAMPLES' document has now be changed to the
'EXAMPLES-AND-INTEGRATION'. It has also been changed to a HTML file
allowing live examples to be demonstrated in the doc.

* Cleaned up some documentation files.

Signed-off-by: staticBanter <staticbanter@staticblogs.ca>