feat: Ability to customize message box classes and ids.
* Due to the recent changes to the configuration file changes to various
`class` and `id` attributes had to be update in the
`site/prod/index.html` file.
* Cleaned up the style file for the application page and examples page.
* Changes to the `createMessageBox` function to allow custom attributes
for message boxes and message back drops. These attributes can be
customized in the `simplePass.config.ts` file.
* Message backdrops can now take up the entire document size.
* Changed how character types are selected in the `simplePass` function.
Rather that using `.pop()` and having to determine if the popped element
is defined we will always select the zeroth element in the array and
then `shift()` our array after we have created a character of that type.
This still needs a check to determine that our array length is greater
than zero defined before we use our array.

Signed-off-by: staticBanter <76642517+staticBanter@users.noreply.github.com>