Rel: 11.0.0-Beta
- `range` function would not return the last number in a list if the
increment was not in step with it. Refactoring the parameters to an
object and inserting a `forceInclusiveEnd` boolean option forces the
last number to be in the list.
- `messageHandler` was reassigning the entire `config` object inside of
the `config.messages` attribute. This has been corrected to only assign
properties that are need for the function.
- Broken link on EXAMPLES-AND-INTEGRATION page.
- Missing whitespace option in list.

- Added Node testing via Mocha and Chai! While this is only the
beginning of all the nessecary testing needed for the application, this
points towards the end of the beta phase of the application.
- Swapping crypto modules can now be done using the
`config.cryptoModule` attribute.
- New options for limiting the size of the repeating characters string
input.
- Two new functions were added:
- `matchCharacterSetConstraint` - Determines if a character code is a
match with a provided character constraint.
- `uniqueCharacterFilter` - Filters a string for only unique
characters.

- `calculateMaxPossibleCharacters` function will now filter out
duplicate characters found in the `excludeCharacters` option.
- `validateModifier` no longer has a `config` parameter.
- `validateModifier` now checks all modifier attributes against a
password modifier schema.
- `validateModifier` now returns true if the validation passed.
- Simplified the logic for returning optional whitespace characters
inside of `generateCharCode`.
- The projects `type` attribute has been set to *'module'* in
`package.json`

- `messageHandler` no longer supports *debug* messages.
- `createModifierList` function was removed (no longer needed).
- `useableAttributes` list has been removed. Replace by using
`Object.keys` on `characterCodeConstraints`. Any attributes that are not
character code constraints have been moved to the `passwordModifier` and
`whitespaceAttributes` lists.
- Old test files and directories such as;
  - test/index.html
  - test/lib

These have been replace with Mocha and Chais CLI packages. (allows for
pivoting to GitHub based testing as-well).

- Due to JSDoc conflicting with with packages of type *module* a blank
`package.json` file was inserted was inserted in the
`docs/site_template` directory. This fools JSDoc and allows us to change
the `publish.cjs` file back to `publish.js`. While this is not
conventional it should not cause any immediate issues and is only
related to the build process of the application.
- Profiling files have been added to the ignores.
- The Webpack `stats` attribute is now true when building the
application in production.
- New NPM script commands:
  - `coverage`. Runs C8 test coverage and produces a report.
  - `test`. Runs the mocha tests.

Signed-off-by: staticBanter <76642517+staticBanter@users.noreply.github.com>